+++ prik/types/__init__.py
@@ -1,1 +1,0 @@
-"""Mappings from semantic types to Python ecosystem types."""